division and branch fields, other minor critical changes
AL-620 HM: Add ‘Division’ and ‘Branch’ fields to the 'Additional details' page
AL-621 The banner text on the actions page when the ticket is in ‘Ready’ state needs to be updated.
AL-631 'Reset changes' button populates wrong data
AL-622 Add a reminder for users to refrain from making any changes to the approved profile during recruitment.